* delete old obsolete classes
